+++ tests-integration/tests/traits_test.rs
@@ -20,6 +20,9 @@ use std::cell::Cell;
 use asyncband::barrier::Barrier;
 use asyncband::broadcast;
 use asyncband::condvar::Condvar;
+use asyncband::event::ManualResetEvent;
+use asyncband::event::ManualResetEventWait;
+use asyncband::event::OwnedManualResetEventWait;
 use asyncband::latch::Latch;
 use asyncband::mpsc;
 use asyncband::mutex::Mutex;
@@ -70,6 +73,9 @@ fn public_types_are_send_and_sync() {
 
     assert_send_and_sync::<Barrier>();
     assert_send_and_sync::<Condvar>();
+    assert_send_and_sync::<ManualResetEvent>();
+    assert_send_and_sync::<ManualResetEventWait<'_>>();
+    assert_send_and_sync::<OwnedManualResetEventWait>();
     assert_send_and_sync::<LazyCell<u32, std::future::Ready<u32>>>();
     assert_send_and_sync::<Once>();
     assert_send_and_sync::<OnceCell<u32>>();
@@ -129,6 +135,9 @@ fn public_types_are_unpin() {
 
     assert_unpin::<Barrier>();
     assert_unpin::<Condvar>();
+    assert_unpin::<ManualResetEvent>();
+    assert_unpin::<ManualResetEventWait<'_>>();
+    assert_unpin::<OwnedManualResetEventWait>();
     assert_unpin::<Latch>();
     assert_unpin::<LazyCell<u32, std::future::Ready<u32>>>();
     assert_unpin::<Once>();